Increase block index type from uint16_t to uint32_t
Expands block index capacity to support more than 65535 blocks by
changing all block index fields and related APIs from uint16_t to uint32_t.

+++ nautilus/src/nautilus/compiler/ir/blocks/BasicBlock.cpp
@@ -7,7 +7,7 @@
 #include <utility>
 
 namespace nautilus::compiler::ir {
-BasicBlock::BasicBlock(uint16_t identifier, std::vector<std::unique_ptr<BasicBlockArgument>>& arguments)
+BasicBlock::BasicBlock(uint32_t identifier, std::vector<std::unique_ptr<BasicBlockArgument>>& arguments)
     : identifier(identifier), operations(), arguments(std::move(arguments)) {
 }
