+++ nautilus/src/nautilus/tracing/Block.cpp
@@ -6,7 +6,7 @@
 
 namespace nautilus::tracing {
 
-Block::Block(uint16_t blockId) : blockId(blockId), type(Type::Default) {
+Block::Block(uint32_t blockId) : blockId(blockId), type(Type::Default) {
 }
 
 operation_identifier Block::addOperation(nautilus::tracing::TraceOperation&& operation) {
